requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
langchain
langchain-community
langchain-google-genai
PyPDF2
faiss-cpu